MDL-66979: Workaround geckodriver clicking on link containing block
There is a bug in Geckodriver which means that it is unable to click any
link which contains a block level node.

See https://github.com/mozilla/geckodriver/issues/653.

The workaround is to click on it's first descendant instead.

This is achieved thorugh finding it's first child and clicking that
instead.

Note: This is not a perfect workaround as the first child element may
not be the one containing the block element, but normally this situation
only occurs where the only immediate child is a block element.